Reverting the AddLayers plugin so adding works.
This change reverts the portion of 968f3d796b86da2bad7ae741506f03d645c5fed7 that changed the AddLayers plugin.  The later changes to the CapabilitiesGrid are not addressed here.  Using the CapabilitiesGrid when layers were added, they weren't being set as the map base layer properly.  It should be easy to to a manual acceptance test using the Bing or MapQuest layer for example.