Finishing the lazy loading implementation for WMS layers.
This gets rid of the forceLazy option. Instead, a list of required layer configuration properties can be provided as requiredProperties option of the source. An empty list here is an equivalent of forceLazy: true.
This changeset also introduces a getState method for layer sources. Applications that persist the application state and have WMS sources without a title configured can now persist a configuration that includes the service title after the Capabilities document is loaded.
Lazy loading no longer prevents other plugins from working, because there are fallbacks for every step in the DescribeLayer - DescribeFeatureType chain. The assumptions made for these fallbacks only work reliably for GeoServer though.