Continue to run if sources fail to load.
Instead of throwing an error when a source is not available, we should be logging errors somewhere accessible to the user.  This change makes it so a viewer can be configured to upload to a layer source and the viewer still works if the layer source is not available (fixes #SUITE-280).