Always do JIT auth challenges.
Previously, the FeatureEditor tools would be disabled by default with no way to enable them if an application didn't have an authentication mechanism (see SUITE-315).  Now we can always enable the tools if there is a schema for the selected layer.  Authentication challenge will come when the user activates a tool.