Usability improvement: confirm text input with Enter key.
Many gxp components instantly apply changes made in forms. For textfields, this is done in a change event listener. The change event triggers when the field blurs, but not when the user hits the Enter key. This can now easily be changed by including a textfield override in the build that blurs the field after a short delay when the user pressed Enter.