Reverting 8e29c18fcc13301953712d9abd33a07bb3385419
Though this might be a nice usability feature, it needs some more work.  When
using a form (FeatureEditorForm), this prevents the user from entering 'd',
using the delete key, or using the backspace key.

See https://github.com/opengeo/mapstory/issues/435

While there are ways of checking the event target that could work, I think it
would be worth revisiting the idea of using the backspace key for deleting
vertices.